requirements.txt
beautifulsoup4>=4.12.3
requests>=2.31.0
pystray>=0.19.5
Pillow>=10.2.0
pyinstaller>=6.4.0
